remove numpy int and float

+++ merlin/util/decoding.py
@@ -322,7 +322,7 @@ class PixelBasedDecoder(object):
         barcodesSeen = np.zeros(self._barcodeCount)
         for b in range(self._barcodeCount):
             barcodeRegions = [x for x in measure.regionprops(
-                        measure.label((decodedImage == b).astype(np.int)))
+                        measure.label((decodedImage == b).astype(np.int32)))
                               if x.area >= self.refactorAreaThreshold]
             barcodesSeen[b] = len(barcodeRegions)
             for br in barcodeRegions:
@@ -360,7 +360,7 @@ class PixelBasedDecoder(object):
         # can be abstracted
         for b in range(self._barcodeCount):
             barcodeRegions = [x for x in measure.regionprops(
-                measure.label((decodedImage == b).astype(np.int)))
+                measure.label((decodedImage == b).astype(np.int32)))
                               if x.area >= 5]
             barcodesSeen[b] = len(barcodeRegions)
             for br in barcodeRegions:
